Fix four bus regulated
Previous Bus1FaultRegulated did not converge.
Now it converges, but generator on reference bus (slack) has not regulators.
For unknown reasons, when the regulated generator is added, the model converges, but the reference node has large powers. Details on document "Check Screens.docx" and package PFTdemo, containing Bus1FaultSI2OK (the regulator model included in the library) and Bus1FaultSI3 (the one with power exchanges in reference node).